Fix release script to not break when meta.json does not exists

+++ scripts/release.ts
@@ -5,6 +5,22 @@ import { join } from "https://deno.land/std@0.181.0/path/mod.ts";
 
 await exec("git fetch --tags");
 
+const exists = async (dir: string): Promise<boolean> => {
+  try {
+    await Deno.stat(dir);
+    // successful, file or directory must exist
+    return true;
+  } catch (error) {
+    if (error instanceof Deno.errors.NotFound) {
+      // file or directory does not exist
+      return false;
+    } else {
+      // unexpected error, maybe permissions, pass it along
+      throw error;
+    }
+  }
+};
+
 const response = await exec(`bash -c "git tag | sort -V | tail -1"`, {
   output: OutputMode.Capture,
 });
@@ -37,33 +53,35 @@ const newVersion = newVersionByUser === "custom"
   : newVersionByUser;
 
 const metaJSONFilePath = join(Deno.cwd(), "meta.json");
-const meta: { version: string } = await Deno.readTextFile(
-  metaJSONFilePath,
-).then(
-  JSON.parse,
-);
-console.log(`Bumping meta.json`);
+if (await exists(metaJSONFilePath)) {
+  const meta: { version: string } = await Deno.readTextFile(
+    metaJSONFilePath,
+  ).then(
+    JSON.parse,
+  );
+  console.log(`Bumping meta.json`);
 
-await Deno.writeTextFile(
-  metaJSONFilePath,
-  JSON.stringify({ ...meta, version: newVersion }, null, 2),
-);
+  await Deno.writeTextFile(
+    metaJSONFilePath,
+    JSON.stringify({ ...meta, version: newVersion }, null, 2),
+  );
 
-const GIT_ADD_COMMAND = `git add ${metaJSONFilePath}`;
-console.log(`Running \`${GIT_ADD_COMMAND}\``);
+  const GIT_ADD_COMMAND = `git add ${metaJSONFilePath}`;
+  console.log(`Running \`${GIT_ADD_COMMAND}\``);
 
-await exec(GIT_ADD_COMMAND);
+  await exec(GIT_ADD_COMMAND);
 
-const GIT_COMMIT_COMMAND =
-  `git commit -m "[live.ts@${newVersion}] Bump meta.json"`;
-console.log(`Running \`${GIT_COMMIT_COMMAND}\``);
+  const GIT_COMMIT_COMMAND =
+    `git commit -m "[live.ts@${newVersion}] Bump meta.json"`;
+  console.log(`Running \`${GIT_COMMIT_COMMAND}\``);
 
-await exec(GIT_COMMIT_COMMAND);
+  await exec(GIT_COMMIT_COMMAND);
 
-const GIT_PUSH_COMMAND = `git push origin main`;
-console.log(`Running \`${GIT_PUSH_COMMAND}\``);
+  const GIT_PUSH_COMMAND = `git push origin main`;
+  console.log(`Running \`${GIT_PUSH_COMMAND}\``);
 
-await exec(GIT_PUSH_COMMAND);
+  await exec(GIT_PUSH_COMMAND);
+}
 
 const GIT_TAG_COMMAND = `git tag ${newVersion}`;
 console.log(`Running \`${GIT_TAG_COMMAND}\``);
